requirements.txt
pytest~=7.4.0
selene~=2.0.0rc4
selenium~=4.11.2
allure-pytest